+++ lua/terrortown/lang/de.lua
@@ -1669,6 +1669,7 @@ L.label_voice_drain_admin = "Entladung pro Tick für Admins und öffentliche Ord
 L.label_voice_drain_recharge = "Aufladungsrate pro Tick wenn nicht gesprochen wird"
 L.label_locational_voice = "Aktiviere Proximity Sprachchat für lebende Spieler"
 L.label_locational_voice_prep = "Aktiviere Proximity Sprachchat während der Vorbereitungszeit"
+L.label_locational_voice_range = "Proximity Sprachchat Reichweite"
 L.label_armor_on_spawn = "Spielerrüstung beim (Neu-)Spawnen"
 L.label_prep_respawn = "Aktiviere automatischen Respawn während der Vorbereitungszeit"
 L.label_preptime_seconds = "Vorbereitungszeit in Sekunden"
@@ -2215,4 +2216,8 @@ L.throw_no_room = "Hier ist kein Platz, um dieses Gerät zu werfen."
 --L.label_crosshair_size_gap = "Crosshair gap size multiplier"
 
 -- 2024-03-31
+--L.help_locational_voice = "Proximity chat is TTT2's implementation of locational 3D voice. Players are only audible in a set radius around them and become quieter the farther away they are."
+--L.help_locational_voice_prep = [[By default the proximity chat is disabled in the preparing phase. Change this convar to also use proximity chat in the preparing phase.
+--
+--Note: Proximity chat is always disabled during the post round phase.]]
 --L.help_voice_duck_spectator = "Ducking spectators makes other spectators quieter in comparison to living players. This can be useful if one wants to listen closely to the discussions of the living players."